+++ DarkCodex/Attributes/PatchInfoAttribute.cs
@@ -33,6 +33,9 @@ namespace DarkCodex
         public bool IsFaulty => (PatchType & Severity.Faulty) > 0;
         public bool IsDangerous => (PatchType & Severity.Create) > 0;
 
+        public string HomebrewStr => Homebrew ? ":house:" : ":book:";
+        public string StatusStr => IsFaulty ? ":x:" : IsWIP ? ":construction:" : "heavy_check_mark";
+
         public int CompareTo(object obj)
         {
             if (this == null && obj == null)