+++ DarkCodex/changelog.md
@@ -3,8 +3,10 @@
 ## [1.1.2]
 - fixed metamagic variants with preferred spell
 - improved ingame menu
+- added one time start up warning
 - added TableTopTweaks' (TTT from now on) Fighter Advanced Weapon Training for limitless shaman weapon
 - added Cursing Gaze | mythic ability: hexes other than grant can be used as a swift action
+- added limitless Skald's Raging Song mythic ability
 
 ## [1.1.1]
 - fixed bleeding attack not counting as bleed for some features